[zsh] Clean up aliases

+++ zsh/modules/aliases.zsh
@@ -8,15 +8,17 @@ alias im="vim"
 
 alias kssh="kitty +kitten ssh"
 
-cpv() {
-    rsync -pogbr -hhh --backup-dir=/tmp/rsync -e /dev/null --progress "$@"
-}
-compdef _files cpv
-alias cpr="rsync -ah --inplace --info=progress2"
-alias rsync-copy="rsync -avz --info=progress2 -h"
-alias rsync-move="rsync -avz --info=progress2 -h --remove-source-files"
-alias rsync-update="rsync -avzu --info=progress2 -h"
-alias rsync-synchronize="rsync -avzu --delete --info=progress2 -h"
+if [ -x "$(command -v rsync)" ]; then
+    cpv() {
+	rsync -pogbr -hhh --backup-dir=/tmp/rsync -e /dev/null --progress "$@"
+    }
+    compdef _files cpv
+    alias cpr="rsync -ah --inplace --info=progress2"
+    alias rsync-copy="rsync -avz --info=progress2 -h"
+    alias rsync-move="rsync -avz --info=progress2 -h --remove-source-files"
+    alias rsync-update="rsync -avzu --info=progress2 -h"
+    alias rsync-synchronize="rsync -avzu --delete --info=progress2 -h"
+fi
 
 certcheck() {
     echo -n | openssl s_client -connect "$@":443 2>&1 | sed -ne '/-BEGIN CERTIFICATE-/,/-END CERTIFICATE-/p' | openssl x509 -text | grep -E "DNS|Subject: CN"
@@ -31,13 +33,17 @@ alias feh="feh --conversion-timeout 1"  # makes feh work with svg
 alias colorcat="pygmentize -g"
 alias diff="diff --color -u"
 alias dd="dd status=progress conv=fsync,fdatasync"
-alias yay="sudo trickle -s -d 6000 yay"
+
+if [ -x "$(command -v trickle)" ]; then
+    alias yay="sudo trickle -s -d 6000 yay"
+fi
+
+if [ -x "$(command -v htop)" ]; then
+    alias htop="btop"
+fi
 
 alias asdf='setxkbmap de neo -option && setxkbmap -option compose:prsc'
 alias uiae='setxkbmap de nodeadkeys -option && setxkbmap -option compose:prsc'
-
-alias please='sudo'
-alias fucking='sudo'
 
 alias pretalx='python -m pretalx'
 alias django='python manage.py'
@@ -65,20 +71,7 @@ function pyclean() {
     find ${ZSH_PYCLEAN_PLACES} -type d -name ".mypy_cache" -delete
 }
 
-##################
-# Suffix Aliases #
-##################
-alias -s de=firefox
-alias -s com=firefox
-
-##################
-# Global Aliases #
-##################
-alias -g !c='| wc -l'
-alias -g !l='| less'
-alias -g !s='| sort'
-alias -g !h='| head'
-alias -g !t='| tail'
-alias -g ...='../..'
-alias -g ....='../../..'
-alias -g .....='../../../..'
+alias ...='../..'
+alias ....='../../..'
+alias .....='../../../..'
+alias ......='../../../../..'
